+++ src/api/dart/mod.rs
@@ -75,28 +75,6 @@ pub use self::{
     },
 };
 
-/// Tries to convert the provided [`DartValueArg`] into a [`MediaSourceKind`].
-///
-/// If the conversion fails, then [`ArgumentError`] is [`return`]ed as a
-/// [`DartFuture`].
-#[macro_export]
-macro_rules! try_into_source_kind {
-    ($k:expr) => {
-        match $k
-            .try_into()
-            .map_err(|err: crate::api::DartValueCastError| {
-                crate::api::ArgumentError::new(
-                    err.value,
-                    "kind",
-                    err.expectation,
-                )
-            }) {
-            Ok(s) => s,
-            Err(e) => return async move { Err(e.into()) }.into_dart_future(),
-        }
-    };
-}
-
 /// Sets the provided [`Dart_Handle`] as a callback for the Rust panic hook.
 #[no_mangle]
 pub unsafe extern "C" fn on_panic(cb: Dart_Handle) {