+++ CHANGELOG.md
@@ -76,6 +76,12 @@ All user visible changes to this project will be documented in this file. This p
     - `LocalMediaInitExceptionKind` variants ([#52]):
         - `GetUserMediaAudioFailed`;
         - `GetUserMediaVideoFailed`.
+    - `RoomHandle.enableRemoteVideo()` and `RoomHandle.disableRemoteVideo()` methods an optional argument to specify `MediaSourceKind` ([#59]).
+
+### Fixed
+
+- Library API:
+    - Unconverted into Dart exception error in `RoomHandle.onFailedLocalMedia()` ([#57]).
 
 ### Updated
 
@@ -90,6 +96,7 @@ All user visible changes to this project will be documented in this file. This p
 [#46]: /../../pull/46
 [#49]: /../../pull/49
 [#52]: /../../pull/52
+[#57]: /../../pull/57